fix: restore MSVC portability and harden timing tests
* fix narrowing conversions reported by MSVC /WX
* keep GCC and Clang warning-clean
* add semantic persistence validation coverage
* restore repository coverage gates
* release v0.19.0-alpha.62

+++ src/services/persistent_state_service.cpp
@@ -158,7 +158,7 @@ void PersistentStateService::begin() {
     }
     constexpr std::size_t kRecordsEnd =
         kCurrentRecordSize + kUserPresetSlotCount * kPresetRecordSize;
-    std::fill_n(storageImage.begin(), kRecordsEnd, 0xFFU);
+    std::fill_n(storageImage.begin(), kRecordsEnd, static_cast<std::uint8_t>(0xFFU));
     if (hasStoredCurrentState_) {
         const auto migrated = serializeCurrentRecord(storedCurrentState_);
         std::copy(migrated.begin(), migrated.end(), storageImage.begin());
